feat: create mocking helpers
